Parser diagnostics - form simplification

+++ src/com/dci/intellij/dbn/diagnostics/action/ParserDiagnosticsDeleteAction.java
@@ -1,6 +1,7 @@
 package com.dci.intellij.dbn.diagnostics.action;
 
 import com.dci.intellij.dbn.common.Icons;
+import com.dci.intellij.dbn.common.util.MessageUtil;
 import com.dci.intellij.dbn.diagnostics.ParserDiagnosticsManager;
 import com.dci.intellij.dbn.diagnostics.data.ParserDiagnosticsResult;
 import com.dci.intellij.dbn.diagnostics.ui.ParserDiagnosticsForm;
@@ -19,11 +20,19 @@ public class ParserDiagnosticsDeleteAction extends AbstractParserDiagnosticsActi
     protected void actionPerformed(@NotNull AnActionEvent e, @NotNull Project project, @NotNull ParserDiagnosticsForm form) {
         ParserDiagnosticsResult result = form.getSelectedResult();
         if (result != null) {
-            ParserDiagnosticsManager manager = getManager(project);
-            manager.deleteResult(result);
-            ParserDiagnosticsResult latestResult = manager.getLatestResult();
-            form.refreshResults();
-            form.selectResult(latestResult);
+            MessageUtil.showQuestionDialog(project,
+                    "Delete diagnostics result",
+                    "Are you sure you want to delete the diagnostic result " + result.getName(),
+                    MessageUtil.OPTIONS_YES_NO, 0,
+                    option -> {
+                        if (option == 0) {
+                            ParserDiagnosticsManager manager = getManager(project);
+                            manager.deleteResult(result);
+                            ParserDiagnosticsResult latestResult = manager.getLatestResult();
+                            form.refreshResults();
+                            form.selectResult(latestResult);
+                        }
+                    });
         }
